+++ backend/database/models.py
@@ -9,6 +9,7 @@ class Authors(models.Model):
     github = models.CharField(max_length = 255, null=True)
     accepted = models.BooleanField(default = False)
     profileImage = models.CharField(max_length = 255)
+    password = models.CharField(max_length = 255)
 
 class Posts(models.Model):
     PLAINTEXT = 'text/plain'
