+++ src/App.css
@@ -1,11 +1,9 @@
 .App {
   text-align: center;
-  display: flex;
+  /* display: flex;
   min-height: 100vh;
   align-items: center;
-  justify-content: center;
-  color: white;
-  background-image: linear-gradient(79deg, #7439db, #C66FBC 48%, #F7944D);
+  justify-content: center; */
 }
 
 .App-logo {
@@ -48,13 +46,27 @@ auth-form-container, .login-form, .register-form {
   flex-direction: column;
 }
 
-@media screen and (min-width: 600px) {
-  .auth-form-container {
-    padding: 5rem;
-    border: 1px solid white;
-    border-radius: 10px;
-    margin: 0.5rem;
-  }
+/* @media screen and (min-width: 600px) {
+  
+} */
+.centered-form{
+  display: flex;
+  min-height: 100vh;
+  align-items: center;
+  justify-content: center;
+}
+
+.register-form-container {
+  padding: 5rem;
+  border: 1px solid black;
+  border-radius: 10px;
+  margin: 0.5rem;
+}
+.login-form-container {
+  padding: 5rem;
+  border: 1px solid black;
+  border-radius: 10px;
+  margin: 0.5rem;
 }
 
 label {
@@ -65,21 +77,21 @@ label {
 input {
   margin: 0.5rem 0;
   padding: 1rem;
-  border: none;
+  border: solid lightblue;
   border-radius: 10px;
 }
 
 button {
   border: none;
-  background-color: white;
+  background-color: lightblue;
   padding: 20px;
   border-radius: 10px;
   cursor: pointer;
-  color: #7439db;
 }
 
 .link-btn {
+  padding: 20px;
+  color: lightblue;
   background: none;
-  color: white;
   text-decoration: underline;
 }
