--- src/components/login/Register.jsx
@@ -1,35 +1,0 @@
-import React, { useState } from "react";
-import axios from 'axios';
-
-
-export const Register = (props) => {
-    const [email, setEmail] = useState('');
-    const [pass, setPass] = useState('');
-    const [name, setName] = useState('');
-
-
-    const handleSubmit = (e) => {
-        e.preventDefault();
-        console.log(email);
-        axios
-            .post('', this.useState)
-            .then(response => {console.log(response)})
-            .catch(error => {console.log(error)})
-    };
-    
-    return (
-        <div className="auth-form-container">
-            <h2>Register</h2>
-        <form className="register-form" onSubmit={handleSubmit}>
-            <label htmlFor="name">Full name</label>
-            <input value={name} onChange={(e) => setName(e.target.value)}name="name" id="name" placeholder="full Name" />
-            <label htmlFor="email">email</label>
-            <input value={email} onChange={(e) => setEmail(e.target.value)}type="email" placeholder="youremail@gmail.com" id="email" name="email" />
-            <label htmlFor="password">password</label>
-            <input value={pass} onChange={(e) => setPass(e.target.value)} type="password" placeholder="********" id="password" name="password" />
-            <button type="submit">Register</button>
-        </form>
-        <button className="link-btn" onClick={() => props.onFormSwitch('login')}>Already have an account? Login here.</button>
-    </div>
-    )
-}
